Fixed bug where reference to Linda object was dropped for a short time ( crashing if GC was run during that time ).
The reference could have also been stored in the module table ( or globals ) but those
can be overwritten by the user. With the Lua registry, it should be inaccessible
by the user and since it's referenced by a unique key, it should also normally be
in accessible to C libraries, making it a relativily safe place to store the reference.