Changed the atexit code to trip the timer thread's write signal.
This fixes one of the causes of the "pthread_cond_destroy(ref) failed, 16 EBUSY" error,
in which the timer thread was stuck waiting on the write_happend signal and never read
the cancel request. This doesn't fix the issue of attempting to cancel threads waiting
on a Linda signal. So this error message ( and seg fault ) can easily be recreated just
by making a new thread, having it wait indefinitly on a Linda and then trying to cancel
the thread.