Fix segfault in LG_lane_new
STACK_END macro was used to check stack size of child Lua state _after_
its thread was launched. That could lead to the check failing as the child
thread started manipulating the stack. The macro then called lua_error
on the child state from the parent thread, causing a segfault as Lua C API
is not thread-safe.

The fix is to place STACK_END statements before THREAD_CREATE statement.